Fix bug SDIO read/write complete early
When used in SBC mode SDIO read/write operations could appear to
complete early. This was caused by task notifications sent by the
SBC module (even though no SBC task was waiting for the notification).

This fix/workaround ensure that we only return when the actual SDIO
operation is complete.